requirements_backup.txt
amadeus==12.0.0
asgiref==3.11.1
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
phonenumbers==9.0.25
pycountry==26.2.16
PyJWT==2.11.0
sqlparse==0.5.5
